refactor: isolate OpenCode tool definitions
