+++ src/tools/index.ts
@@ -1,372 +1,1 @@
-import { tool, type ToolDefinition } from "@opencode-ai/plugin";
-
-import type { HostMode } from "../config/host.js";
-import type { ParsedCodebaseIndexConfig } from "../config/schema.js";
-import type { Indexer } from "../indexer/index.js";
-import {
-  CALL_GRAPH_DIRECTIONS,
-  CHUNK_TYPES,
-  INDEX_LOG_LEVELS,
-  INDEX_LOG_CATEGORIES,
-  RELATIONSHIP_TYPES,
-} from "./contracts.js";
-import {
-  DEFAULT_CONTEXT_PACK_TOKEN_BUDGET,
-  formatCodebasePeek,
-  formatSearchResults,
-  MAX_CONTEXT_PACK_TOKEN_BUDGET,
-  MIN_CONTEXT_PACK_TOKEN_BUDGET,
-} from "./utils.js";
-import {
-  addKnowledgeBase,
-  findSimilarCode,
-  getIndexerForProject as getOperationIndexerForProject,
-  getSharedIndexer as getOperationSharedIndexer,
-  initializeTools as initializeToolOperations,
-  listKnowledgeBases,
-  removeKnowledgeBase,
-  searchCodebaseWithEffectiveness,
-} from "./operations.js";
-import { pr_impact } from "./pr-impact.js";
-import {
-  executeCallGraph,
-  executeCallGraphPath,
-  executeCodebaseContext,
-  executeIndexCodebase,
-  executeIndexHealthCheck,
-  executeIndexLogs,
-  executeIndexMetrics,
-  executeIndexStatus,
-  executeImplementationLookup,
-} from "./execute-common.js";
-import {
-  MAX_CONTEXT_PATH_DEPTH,
-  MAX_CONTEXT_RESULT_LIMIT,
-  MIN_CONTEXT_PATH_DEPTH,
-  MIN_CONTEXT_RESULT_LIMIT,
-} from "./context.js";
-import { writeFileSync } from "fs";
-import * as os from "os";
-import * as path from "path";
-import { attachRecentActivity } from "./visualize/activity.js";
-import { generateVisualizationHtml, transformForVisualization } from "./visualize/index.js";
-
-const z = tool.schema;
-const DEFAULT_HOST: HostMode = "opencode";
-const CHUNK_TYPE_VALUES = CHUNK_TYPES;
-const RELATIONSHIP_TYPE_VALUES = RELATIONSHIP_TYPES;
-
-export function initializeTools(projectRoot: string, config: ParsedCodebaseIndexConfig): void {
-  initializeToolOperations(projectRoot, config, DEFAULT_HOST);
-}
-
-export function getSharedIndexer(): Indexer {
-  return getOperationSharedIndexer(DEFAULT_HOST);
-}
-
-export function getIndexerForProject(directory: string): Indexer {
-  return getOperationIndexerForProject(directory, DEFAULT_HOST);
-}
-
-export const codebase_context: ToolDefinition = tool({
-  description:
-    "PREFERRED FIRST TOOL for repository questions. Returns a deduplicated, file-diverse evidence pack within tokenBudget. Provide from and to for dependency paths, with optional fromFilePath/toFilePath when names are ambiguous; provide symbol for definitions; or provide only query for conceptual discovery.",
-  args: {
-    query: z.string().describe("The repository question or behavior to locate"),
-    from: z.string().nullable().optional().describe("Source symbol for a dependency path"),
-    to: z.string().nullable().optional().describe("Target symbol for a dependency path"),
-    fromFilePath: z.string().nullable().optional().describe("Optional source file path used only to disambiguate duplicate source names"),
-    toFilePath: z.string().nullable().optional().describe("Optional target file path used only to disambiguate duplicate target names"),
-    symbol: z.string().nullable().optional().describe("Exact symbol for authoritative definition lookup"),
-    limit: z.number().int().min(MIN_CONTEXT_RESULT_LIMIT).max(MAX_CONTEXT_RESULT_LIMIT).nullable().optional().default(10)
-      .describe(`Maximum number of results (${MIN_CONTEXT_RESULT_LIMIT}-${MAX_CONTEXT_RESULT_LIMIT})`),
-    maxDepth: z.number().int().min(MIN_CONTEXT_PATH_DEPTH).max(MAX_CONTEXT_PATH_DEPTH).nullable().optional().default(10)
-      .describe(`Maximum call-path traversal depth (${MIN_CONTEXT_PATH_DEPTH}-${MAX_CONTEXT_PATH_DEPTH})`),
-    fileType: z.string().nullable().optional().describe("Filter by file extension"),
-    directory: z.string().nullable().optional().describe("Filter by directory path"),
-    tokenBudget: z.number().int().min(MIN_CONTEXT_PACK_TOKEN_BUDGET).max(MAX_CONTEXT_PACK_TOKEN_BUDGET)
-      .nullable().optional().default(DEFAULT_CONTEXT_PACK_TOKEN_BUDGET)
-      .describe(`Maximum response tokens (${MIN_CONTEXT_PACK_TOKEN_BUDGET}-${MAX_CONTEXT_PACK_TOKEN_BUDGET})`),
-  },
-  async execute(args, context) {
-    return (await executeCodebaseContext(context?.worktree, DEFAULT_HOST, args)).text;
-  },
-});
-
-export const codebase_peek: ToolDefinition = tool({
-  description:
-    "Quick lookup of code locations by meaning. Returns only metadata (file, line, name, type) WITHOUT code content. Use this first to find WHERE code is, then use Read tool to examine specific files. Saves tokens by not returning full code blocks. Best for: discovery, navigation, finding multiple related locations.",
-  args: {
-    query: z.string().describe("Natural language description of what code you're looking for."),
-    limit: z.number().optional().default(10).describe("Maximum number of results to return"),
-    fileType: z.string().optional().describe("Filter by file extension (e.g., 'ts', 'py', 'rs')"),
-    directory: z.string().optional().describe("Filter by directory path (e.g., 'src/utils', 'lib')"),
-    chunkType: z.enum(CHUNK_TYPE_VALUES).optional().describe("Filter by code chunk type"),
-    blameAuthor: z.string().optional().describe("Filter by git blame author name or email"),
-    blameSha: z.string().optional().describe("Filter by git blame commit SHA or prefix"),
-    blameSince: z.string().optional().describe("Filter to chunks last changed on or after this date (e.g., 2025-01-01)"),
-  },
-  async execute(args, context) {
-    return searchCodebaseWithEffectiveness(context?.worktree, DEFAULT_HOST, "peek", args.query, {
-      limit: args.limit ?? 10,
-      fileType: args.fileType,
-      directory: args.directory,
-      chunkType: args.chunkType,
-      metadataOnly: true,
-      blameAuthor: args.blameAuthor,
-      blameSha: args.blameSha,
-      blameSince: args.blameSince,
-    }, (results) => {
-      const text = formatCodebasePeek(results);
-      return { output: text, text };
-    });
-  },
-});
-
-export const index_codebase: ToolDefinition = tool({
-  description:
-    "Index the codebase for semantic search. Creates vector embeddings of code chunks. Incremental - only re-indexes changed files (~50ms when nothing changed). Run before first codebase_search.",
-  args: {
-    force: z.boolean().optional().default(false).describe("Force reindex even if already indexed"),
-    estimateOnly: z.boolean().optional().default(false).describe("Only show cost estimate without indexing"),
-    verbose: z.boolean().optional().default(false).describe("Show detailed info about skipped files and parsing failures"),
-  },
-  async execute(args, context) {
-    return (await executeIndexCodebase(context?.worktree, DEFAULT_HOST, args, (title, metadata) => {
-      context.metadata({ title, metadata });
-    })).text;
-  },
-});
-
-export const index_status: ToolDefinition = tool({
-  description:
-    "Check the status of the codebase index. Shows whether the codebase is indexed, how many chunks are stored, and the embedding provider being used.",
-  args: {},
-  async execute(_args, context) {
-    return (await executeIndexStatus(context?.worktree, DEFAULT_HOST)).text;
-  },
-});
-
-export const index_health_check: ToolDefinition = tool({
-  description:
-    "Check index health and remove stale entries from deleted files. Run this to clean up the index after files have been deleted.",
-  args: {},
-  async execute(_args, context) {
-    return (await executeIndexHealthCheck(context?.worktree, DEFAULT_HOST)).text;
-  },
-});
-
-export const index_metrics: ToolDefinition = tool({
-  description:
-    "Get operational metrics plus opt-in privacy-safe repository-tool effectiveness counters. Use reset=true to clear in-memory metrics before reading. Operational metrics require debug.enabled=true and debug.metrics=true. Privacy-safe aggregates require only effectivenessMetrics.enabled=true.",
-  args: {
-    reset: z.boolean().optional().default(false).describe("Reset in-memory operational and effectiveness metrics before returning the snapshot"),
-  },
-  async execute(args, context) {
-    return (await executeIndexMetrics(context?.worktree, DEFAULT_HOST, args)).text;
-  },
-});
-
-export const index_logs: ToolDefinition = tool({
-  description:
-    "Get recent debug logs from the codebase indexer. Shows timestamped log entries with level and category. Requires debug.enabled=true in config.",
-  args: {
-    limit: z.number().optional().default(20).describe("Maximum number of log entries to return"),
-    category: z.enum(INDEX_LOG_CATEGORIES).optional().describe("Filter by log category"),
-    level: z.enum(INDEX_LOG_LEVELS).optional().describe("Filter by minimum log level"),
-  },
-  async execute(args, context) {
-    return (await executeIndexLogs(context?.worktree, DEFAULT_HOST, args)).text;
-  },
-});
-
-export const find_similar: ToolDefinition = tool({
-  description:
-    "Find code similar to a given snippet. Use for duplicate detection, pattern discovery, or refactoring prep. Paste code and find semantically similar implementations elsewhere in the codebase.",
-  args: {
-    code: z.string().describe("The code snippet to find similar code for"),
-    limit: z.number().optional().default(10).describe("Maximum number of results to return"),
-    fileType: z.string().optional().describe("Filter by file extension (e.g., 'ts', 'py', 'rs')"),
-    directory: z.string().optional().describe("Filter by directory path (e.g., 'src/utils', 'lib')"),
-    chunkType: z.enum(CHUNK_TYPE_VALUES).optional().describe("Filter by code chunk type"),
-    excludeFile: z.string().optional().describe("Exclude results from this file path (useful when searching for duplicates of code from a specific file)"),
-  },
-  async execute(args, context) {
-    const results = await findSimilarCode(context?.worktree, DEFAULT_HOST, args.code, {
-      limit: args.limit,
-      fileType: args.fileType,
-      directory: args.directory,
-      chunkType: args.chunkType,
-      excludeFile: args.excludeFile,
-    });
-
-    if (results.length === 0) {
-      return "No similar code found. Try a different snippet or run index_codebase first.";
-    }
-
-    return formatSearchResults(results);
-  },
-});
-
-export const codebase_search: ToolDefinition = tool({
-  description:
-    "Search codebase by MEANING, not keywords. Returns full code content. Use when you need to see actual implementation. For just finding WHERE code is (saves ~90% tokens), use codebase_peek instead. For known identifiers like 'validateToken', use grep - it's faster.",
-  args: {
-    query: z.string().describe("Natural language description of what code you're looking for. Describe behavior, not syntax."),
-    limit: z.number().optional().default(5).describe("Maximum number of results to return"),
-    fileType: z.string().optional().describe("Filter by file extension (e.g., 'ts', 'py', 'rs')"),
-    directory: z.string().optional().describe("Filter by directory path (e.g., 'src/utils', 'lib')"),
-    chunkType: z.enum(CHUNK_TYPE_VALUES).optional().describe("Filter by code chunk type"),
-    contextLines: z.number().optional().describe("Number of extra lines to include before/after each match (default: 0)"),
-    blameAuthor: z.string().optional().describe("Filter by git blame author name or email"),
-    blameSha: z.string().optional().describe("Filter by git blame commit SHA or prefix"),
-    blameSince: z.string().optional().describe("Filter to chunks last changed on or after this date (e.g., 2025-01-01)"),
-  },
-  async execute(args, context) {
-    return searchCodebaseWithEffectiveness(context?.worktree, DEFAULT_HOST, "search", args.query, {
-      limit: args.limit ?? 5,
-      fileType: args.fileType,
-      directory: args.directory,
-      chunkType: args.chunkType,
-      contextLines: args.contextLines,
-      blameAuthor: args.blameAuthor,
-      blameSha: args.blameSha,
-      blameSince: args.blameSince,
-    }, (results) => {
-      const text = results.length === 0
-        ? "No matching code found. Try a different query or run index_codebase first."
-        : formatSearchResults(results, "score");
-      return { output: text, text };
-    });
-  },
-});
-
-export const implementation_lookup: ToolDefinition = tool({
-  description:
-    "Jump to symbol definition. Find WHERE something is defined. " +
-    "Returns the authoritative source location(s) for a function, class, method, type, or variable. " +
-    "Prefers real implementation files over tests, docs, examples, and fixtures. " +
-    "Use when you need the definition site, not all usages.",
-  args: {
-    query: z.string().describe("Symbol name or natural language description (e.g., 'validateToken', 'where is the payment handler defined')"),
-    limit: z.number().optional().default(5).describe("Maximum number of results"),
-    fileType: z.string().optional().describe("Filter by file extension (e.g., 'ts', 'py')"),
-    directory: z.string().optional().describe("Filter by directory path (e.g., 'src/utils')"),
-  },
-  async execute(args, context) {
-    return (await executeImplementationLookup(context?.worktree, DEFAULT_HOST, args)).text;
-  },
-});
-
-export const call_graph: ToolDefinition = tool({
-  description:
-    "Query the call graph by function or method name to find direct callers or callees. Unique names resolve automatically; use filePath only when duplicate names are reported."
-    + " Supports relationship types: Call, MethodCall, Constructor, Import, Inherits, Implements.",
-  args: {
-    name: z.string().describe("Function or method name to query"),
-    direction: z.enum(CALL_GRAPH_DIRECTIONS).default("callers").describe("Direction: 'callers' finds who calls this function, 'callees' finds what this function calls"),
-    filePath: z.string().optional().describe("Optional file path used to disambiguate duplicate symbol names"),
-    symbolId: z.string().optional().describe("Optional backward-compatible symbol ID escape hatch"),
-    relationshipType: z.enum(RELATIONSHIP_TYPE_VALUES).optional().describe("Filter by relationship type. Omit to show all."),
-  },
-  async execute(args, context) {
-    return (await executeCallGraph(context?.worktree, DEFAULT_HOST, args)).text;
-  },
-});
-
-export const call_graph_path: ToolDefinition = tool({
-  description:
-    "Find the shortest connection path between two named symbols. Unique names resolve automatically; use fromFilePath or toFilePath only when duplicate endpoint names are reported.",
-  args: {
-    from: z.string().describe("Source function/method name (starting point)"),
-    to: z.string().describe("Target function/method name (destination)"),
-    fromFilePath: z.string().optional().describe("Optional source file path used to disambiguate duplicate source names"),
-    toFilePath: z.string().optional().describe("Optional target file path used to disambiguate duplicate target names"),
-    maxDepth: z.number().optional().default(10).describe("Maximum traversal depth (default: 10)"),
-  },
-  async execute(args, context) {
-    return (await executeCallGraphPath(context?.worktree, DEFAULT_HOST, args)).text;
-  },
-});
-
-export const add_knowledge_base: ToolDefinition = tool({
-  description:
-    "Add a folder as a knowledge base to the semantic search index. " +
-    "The folder will be indexed alongside the main project code. " +
-    "Supports absolute paths or relative paths (relative to the project root).",
-  args: {
-    path: z.string().describe("Path to the folder to add as a knowledge base (absolute or relative to the project root)"),
-  },
-  async execute(args, context) {
-    return addKnowledgeBase(context?.worktree, DEFAULT_HOST, args.path);
-  },
-});
-
-export const list_knowledge_bases: ToolDefinition = tool({
-  description:
-    "List all configured knowledge base folders that are indexed alongside the main project.",
-  args: {},
-  async execute(_args, context) {
-    return listKnowledgeBases(context?.worktree, DEFAULT_HOST);
-  },
-});
-
-export const remove_knowledge_base: ToolDefinition = tool({
-  description:
-    "Remove a knowledge base folder from the semantic search index.",
-  args: {
-    path: z.string().describe("Path of the knowledge base to remove (must match the configured path exactly)"),
-  },
-  async execute(args, context) {
-    return removeKnowledgeBase(context?.worktree, DEFAULT_HOST, args.path.trim());
-  },
-});
-
-export { pr_impact };
-
-export const index_visualize: ToolDefinition = tool({
-  description:
-    "Generate an interactive HTML visualization of recent code movement and the call graph. " +
-    "Starts with temporal onboarding context from Git history, then supports module, symbol, hotspot, and cycle drill-down.",
-  args: {
-    directory: z.string().optional().describe("Filter to symbols in this directory (e.g., 'src/services')"),
-    maxNodes: z.number().optional().default(5000).describe("Maximum nodes to include (default 5000)"),
-    includeOrphans: z.boolean().optional().default(false).describe("Include symbols with no call relationships"),
-  },
-  async execute(args, context) {
-    const projectRoot = context?.worktree ?? process.cwd();
-    const indexer = getIndexerForProject(projectRoot);
-    const rawData = await indexer.getVisualizationData({
-      directory: args.directory,
-    });
-
-    if (rawData.symbols.length === 0) {
-      return "No call graph data found. Run index_codebase first to build the call graph.";
-    }
-
-    const vizData = attachRecentActivity(transformForVisualization(rawData.symbols, rawData.edges, {
-      includeOrphans: args.includeOrphans,
-      directory: args.directory,
-      maxNodes: args.maxNodes,
-    }), projectRoot);
-
-    if (vizData.nodes.length === 0) {
-      return "No connected symbols found for visualization. Try including orphans with includeOrphans=true, or check that the call graph has resolved edges.";
-    }
-
-    const html = generateVisualizationHtml(vizData);
-    const outputPath = path.join(os.tmpdir(), `call-graph-${Date.now()}.html`);
-    writeFileSync(outputPath, html, "utf-8");
-
-    let result = `Temporal call graph visualization generated: ${outputPath}\n\n`;
-    result += `Nodes: ${vizData.nodes.length} | Edges: ${vizData.edges.length}\n`;
-    result += `Recent change lenses: ${vizData.changes?.length ?? 0}\n`;
-    result += `Files: ${new Set(vizData.nodes.map(n => n.filePath)).size}\n`;
-    result += `Directories: ${new Set(vizData.nodes.map(n => n.directory)).size}`;
-    if (vizData.metadata.truncated) {
-      result += `\n\n\u26a0\ufe0f Graph truncated to ${args.maxNodes} most-connected nodes (total: ${rawData.symbols.length}).`;
-    }
-
-    return result;
-  },
-});
+export * from "../adapters/opencode/tools.js";
